requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
autoflake==1.4
bcrypt==3.2.2
black==22.6.0
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
distlib==0.3.5
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
fastapi-users==10.1.5
fastapi-users-db-sqlalchemy==4.0.3
filelock==3.7.1
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
identify==2.5.2
idna==3.3
makefun==1.14.0
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.7.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
PyJWT==2.4.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
virtualenv==20.16.2
